requirements.txt
# =========================

# ☕ Coffee Shop Backend - requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
django-cors-headers>=4.3
python-dotenv>=1.0
Pillow>=10.0
django-filter>=23.0
gunicorn>=21.2
redis>=5.0
celery>=5.3
drf-yasg>=1.21
bcrypt>=4.0
django-debug-toolbar>=4.2
